[Fix] : 채팅메시지 MVVM 으로 livedata로 실시간채팅으로
