+++ app/build.gradle
@@ -3,6 +3,8 @@ plugins {
     id 'org.jetbrains.kotlin.android'
     id 'com.google.gms.google-services'
     id 'androidx.navigation.safeargs.kotlin'
+    id 'kotlin-parcelize'//직렬화로추가
+    id 'kotlin-android'//직렬화로추가
 }
 
 android {
@@ -47,6 +49,10 @@ dependencies {
     implementation 'de.hdodenhof:circleimageview:3.1.0'//써클이미지뷰 라이브러리
     implementation "androidx.viewpager2:viewpager2:1.0.0" //뷰페이저
     implementation 'com.tbuonomo:dotsindicator:4.2' //닷츠 인디케이터 라이브러러리
+    implementation 'com.github.bumptech.glide:glide:4.12.0' //글라이드
+
+    implementation "androidx.fragment:fragment-ktx:1.3.6"
+    implementation "com.airbnb.android:lottie:6.1.0" //lottie
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
