+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,12 @@
         android:theme="@style/Theme.FromPet"
         tools:targetApi="31">
         <activity
+            android:name=".chating.ChatUserDetailActivity"
+            android:exported="false" />
+        <activity
+            android:name=".chating.ChatMessageActivity"
+            android:exported="false" />
+        <activity
             android:name=".chating.ChatingActivity"
             android:exported="false" />
         <activity
