+++ app/src/main/java/com/example/frompet/chating/ChatMessageActivity.kt
@@ -3,13 +3,13 @@ package com.example.frompet.chating
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import androidx.activity.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.frompet.R
-import com.example.frompet.chating.adapter.ChatListAdapter
 import com.example.frompet.chating.adapter.ChatMessageAdapter
 import com.example.frompet.databinding.ActivityChatMessageBinding
 import com.example.frompet.login.data.ChatMessage
 import com.example.frompet.login.data.UserModel
+import com.example.frompet.login.viewmodel.ChatViewModel
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.database.FirebaseDatabase
@@ -19,77 +19,35 @@ import com.google.firebase.database.ValueEventListener
 
 class ChatMessageActivity : AppCompatActivity() {
     private lateinit var binding: ActivityChatMessageBinding
+    private val viewModel: ChatViewModel by viewModels()
     private val adapter = ChatMessageAdapter(FirebaseAuth.getInstance().currentUser?.uid ?: "")
-    private val database = FirebaseDatabase.getInstance().reference
-    private val firestore = FirebaseFirestore.getInstance()
-    private val auth = FirebaseAuth.getInstance()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         binding = ActivityChatMessageBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         binding.rvMessage.adapter = adapter
         binding.rvMessage.layoutManager = LinearLayoutManager(this)
 
+        viewModel.chatMessages.observe(this) { messages ->
+            adapter.submitList(messages) {
+                binding.rvMessage.post {//맨마지막채팅이 맨 아래에 오게하긔
+                    binding.rvMessage.scrollToPosition(messages.size - 1)
+                }
+            }
+        }
+
         val user: UserModel? = intent.getParcelableExtra("user")
         user?.let {
             binding.ivSendBtn.setOnClickListener {
                 val message = binding.etMessage.text.toString()
                 if (message.isNotEmpty()) {
-                    sendMessage(user.uid, message)
+                    viewModel.sendMessage(user.uid, message)
+                    binding.etMessage.text.clear()
                 }
             }
         }
-        loadPreviousMessages()
+        viewModel.loadPreviousMessages()
     }
-
-    private fun sendMessage(receiverId: String, message: String) {
-        val currentUserId = FirebaseAuth.getInstance().currentUser?.uid ?: return
-
-        firestore.collection("User").document(currentUserId).get()
-            .addOnSuccessListener { document ->
-                val currentUser = document.toObject(UserModel::class.java)
-                val senderPetName = currentUser?.petName ?: "Unknown"
-
-                val chatMessage = ChatMessage(
-                    senderId = currentUserId,
-                    senderPetName = senderPetName,
-                    receiverId = receiverId,
-                    message = message,
-                    timestamp = System.currentTimeMillis()
-                )
-
-                database.child("chatMessages").push().setValue(chatMessage)
-                    .addOnSuccessListener {
-                        binding.etMessage.text.clear()
-                        adapter.submitList(adapter.currentList + chatMessage)
-                    }
-                    .addOnFailureListener { exception ->
-                        Log.d("jun", "메시지전송실패: ${exception.message}")
-                    }
-            }
-            .addOnFailureListener { exception ->
-                Log.d("jun", "연결실패: ${exception.message}")
-            }
-    }
-
-    private fun loadPreviousMessages() {
-        val currentUserId = auth.currentUser?.uid ?: return
-        //박세준
-        // 현재 사용자가 관련된 모든 메시지를 불러오기 위한 쿼리입니다
-        database.child("chatMessages").addListenerForSingleValueEvent(object : ValueEventListener {
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                val allMessages = dataSnapshot.children.mapNotNull { it.getValue(ChatMessage::class.java) }
-                    .filter { it.senderId == currentUserId || it.receiverId == currentUserId }
-                adapter.submitList(allMessages.sortedBy { it.timestamp })
-            }
-
-            override fun onCancelled(databaseError: DatabaseError) {
-                Log.d("jun", "불러오기실패: ${databaseError.message}")
-            }
-        })
-    }
-
 }
